Restructure package and rename main
