+++ Dockerfile
@@ -3,4 +3,4 @@ WORKDIR /app
 COPY . .
 RUN pip install --no-cache-dir -r requirements.txt
 RUN pip install --no-cache-dir pandas-ta==0.3.14b0 --no-binary :all:
-CMD ["python", "main.py"]
+CMD ["python", "-m", "finansal_analiz_sistemi"]
