+++ finansal_analiz_sistemi/__main__.py
@@ -1,0 +1,3 @@
+import runpy
+
+runpy.run_module("run", run_name="__main__")
