+++ setup.py
@@ -27,14 +27,13 @@ setup(
     py_modules=[
         "backtest_core",
         "config",
-        "data_loader",
         "data_loader_cache",
         "filter_engine",
         "filtre_dogrulama",
         "indicator_calculator",
         "kontrol_araci",
         "log_to_health",
-        "main",
+        "run",
         "preprocessor",
         "report_generator",
         "report_stats",
